requirements.txt
discord.py>=2.3.0
Flask>=2.0.0
aiohttp>=3.8.0
python-dotenv